ml_dml_generator/requirements.txt

flask
python-docx
transformers
torch
